[GREEN] Implement JWT Service token deletion

+++ src/main/java/id/ac/ui/cs/advprog/bookku/bookku_userauth/service/JWTService.java
@@ -49,7 +49,7 @@ public class JWTService {
 
         return Jwts.builder()
                 .setClaims(claims)
-                .setSubject(Integer.toString(account.getId()))
+                .setSubject(Long.toString(account.getId()))
                 .setIssuedAt(currentDate)
                 .setExpiration(expirationTime)
                 .signWith(getSignInKey(), SignatureAlgorithm.HS256)
